Moves generic testing to inheritance based framework, refactors test for BaseTransformer and DataFrameMethodTransformer
