+++ tests/base/test_BaseTransformer.py
@@ -2,10 +2,36 @@ import pytest
 import test_aide as ta
 
 import tests.test_data as d
+from tests.base_tests import (
+    GenericFitTests,
+    GenericTransformTests,
+    OtherBaseBehaviourTests,
+)
+from tests.specific_column_type_tests import ColumnStrListInitTests
 from tubular.base import BaseTransformer
 
 
-class TestTransform:
+class TestInit(ColumnStrListInitTests):
+    """Generic tests for transformer.init()."""
+
+    @classmethod
+    def setup_class(cls):
+        cls.transformer_name = "BaseTransformer"
+
+
+class TestFit(GenericFitTests):
+    """Generic tests for transformer.fit()"""
+
+    @classmethod
+    def setup_class(cls):
+        cls.transformer_name = "BaseTransformer"
+
+
+class TestTransform(GenericTransformTests):
+    @classmethod
+    def setup_class(cls):
+        cls.transformer_name = "BaseTransformer"
+
     @pytest.mark.parametrize(
         ("df", "expected"),
         ta.pandas.adjusted_dataframe_params(d.create_df_1(), d.create_df_1()),
@@ -21,3 +47,9 @@ class TestTransform:
             actual=df_transformed,
             msg="Check X returned from transform",
         )
+
+
+class TestOtherBaseBehaviour(OtherBaseBehaviourTests):
+    @classmethod
+    def setup_class(cls):
+        cls.transformer_name = "BaseTransformer"
