+++ tests/base_tests.py
@@ -1,0 +1,450 @@
+# tests to apply to all columns str or list transformers
+import re
+
+import numpy as np
+import pandas as pd
+import pytest
+import sklearn.base as b
+
+import tests.test_data as d
+
+
+class GenericInitTests:
+    """
+    Generic tests for transformer.init().
+    Note this deliberately avoids starting with "Tests" so that the tests are not run on import.
+    """
+
+    def test_print(self, instantiated_transformers):
+        """Test that transformer can be printed.
+        If an error is raised in this test it will not prevent the transformer from working correctly,
+        but will stop other unit tests passing.
+        """
+
+        print(instantiated_transformers[self.transformer_name])
+
+    def test_clone(self, instantiated_transformers):
+        """Test that transformer can be used in sklearn.base.clone function."""
+
+        b.clone(instantiated_transformers[self.transformer_name])
+
+    @pytest.mark.parametrize("non_bool", [1, "True", {"a": 1}, [1, 2], None])
+    def test_verbose_non_bool_error(
+        self,
+        non_bool,
+        minimal_attribute_dict,
+        uninstantiated_transformers,
+    ):
+        """Test an error is raised if verbose is not specified as a bool."""
+
+        with pytest.raises(
+            TypeError,
+            match=f"{self.transformer_name}: verbose must be a bool",
+        ):
+            uninstantiated_transformers[self.transformer_name](
+                verbose=non_bool,
+                **minimal_attribute_dict[self.transformer_name],
+            )
+
+    @pytest.mark.parametrize("non_bool", [1, "True", {"a": 1}, [1, 2], None])
+    def test_copy_non_bool_error(
+        self,
+        non_bool,
+        minimal_attribute_dict,
+        uninstantiated_transformers,
+    ):
+        """Test an error is raised if copy is not specified as a bool."""
+
+        with pytest.raises(
+            TypeError,
+            match=f"{self.transformer_name}: copy must be a bool",
+        ):
+            uninstantiated_transformers[self.transformer_name](
+                copy=non_bool,
+                **minimal_attribute_dict[self.transformer_name],
+            )
+
+    @pytest.mark.parametrize("non_string_or_list", [1, True, {"a": 1}, None])
+    def test_columns_non_string_error(
+        self,
+        non_string_or_list,
+        minimal_attribute_dict,
+        uninstantiated_transformers,
+    ):
+        """Test an error is raised if columns is not passed as a string or list."""
+
+        args = minimal_attribute_dict[self.transformer_name].copy()
+        args["columns"] = non_string_or_list
+
+        with pytest.raises(
+            TypeError,
+            match=re.escape(
+                f"{self.transformer_name}: columns must be a string or list with the columns to be pre-processed (if specified)",
+            ),
+        ):
+            uninstantiated_transformers[self.transformer_name](**args)
+
+
+class GenericFitTests:
+    """
+    Generic tests for transformer.fit().
+    Note this deliberately avoids starting with "Tests" so that the tests are not run on import.
+    """
+
+    def test_fit_returns_self(
+        self,
+        instantiated_transformers,
+    ):
+        """Test fit returns self?."""
+
+        df = d.create_numeric_df_1()
+
+        x = instantiated_transformers[self.transformer_name]
+
+        x_fitted = x.fit(df, df["c"])
+
+        assert (
+            x_fitted is x
+        ), f"Returned value from {self.transformer_name}.fit not as expected."
+
+    @pytest.mark.parametrize("non_df", [1, True, "a", [1, 2], {"a": 1}, None])
+    def test_X_non_df_error(
+        self,
+        instantiated_transformers,
+        non_df,
+    ):
+        """Test an error is raised if X is not passed as a pd.DataFrame."""
+
+        df = d.create_numeric_df_1()
+
+        x = instantiated_transformers[self.transformer_name]
+
+        with pytest.raises(
+            TypeError,
+            match=f"{self.transformer_name}: X should be a pd.DataFrame",
+        ):
+            x.fit(non_df, df["a"])
+
+    @pytest.mark.parametrize("non_series", [1, True, "a", [1, 2], {"a": 1}])
+    def test_non_pd_type_error(
+        self,
+        non_series,
+        instantiated_transformers,
+    ):
+        """Test an error is raised if y is not passed as a pd.Series."""
+
+        df = d.create_numeric_df_1()
+
+        x = instantiated_transformers[self.transformer_name]
+
+        with pytest.raises(
+            TypeError,
+            match=f"{self.transformer_name}: unexpected type for y, should be a pd.Series",
+        ):
+            x.fit(X=df, y=non_series)
+
+    def test_X_no_rows_error(
+        self,
+        instantiated_transformers,
+    ):
+        """Test an error is raised if X has no rows."""
+
+        x = instantiated_transformers[self.transformer_name]
+
+        df = pd.DataFrame(columns=["a", "b", "c"])
+
+        with pytest.raises(
+            ValueError,
+            match=re.escape(f"{self.transformer_name}: X has no rows; (0, 3)"),
+        ):
+            x.fit(df, df["a"])
+
+    def test_Y_no_rows_error(
+        self,
+        instantiated_transformers,
+    ):
+        """Test an error is raised if Y has no rows."""
+
+        x = instantiated_transformers[self.transformer_name]
+
+        df = pd.DataFrame({"a": 1, "b": "wow", "c": np.nan}, index=[0])
+
+        with pytest.raises(
+            ValueError,
+            match=re.escape(f"{self.transformer_name}: y is empty; (0,)"),
+        ):
+            x.fit(X=df, y=pd.Series(name="d", dtype=object))
+
+    def test_unexpected_kwarg_error(
+        self,
+        minimal_attribute_dict,
+        uninstantiated_transformers,
+    ):
+        with pytest.raises(
+            TypeError,
+            match=re.escape(
+                "__init__() got an unexpected keyword argument 'unexpected_kwarg'",
+            ),
+        ):
+            uninstantiated_transformers[self.transformer_name](
+                unexpected_kwarg="spanish inquisition",
+                **minimal_attribute_dict[self.transformer_name],
+            )
+
+
+class GenericTransformTests:
+    """
+    Generic tests for transformer.fit().
+    Note this deliberately avoids starting with "Tests" so that the tests are not run on import.
+    """
+
+    @pytest.mark.parametrize("non_df", [1, True, "a", [1, 2], {"a": 1}, None])
+    def test_non_pd_type_error(
+        self,
+        non_df,
+        instantiated_transformers,
+    ):
+        """Test that an error is raised in transform is X is not a pd.DataFrame."""
+
+        df = d.create_df_10()
+
+        x = instantiated_transformers[self.transformer_name]
+
+        x_fitted = x.fit(df, df["c"])
+
+        with pytest.raises(
+            TypeError,
+            match=f"{self.transformer_name}: X should be a pd.DataFrame",
+        ):
+            x_fitted.transform(X=non_df)
+
+    def test_copy_returned(self, minimal_attribute_dict, uninstantiated_transformers):
+        """Test check that a copy is returned if copy is set to True"""
+        df = d.create_df_10()
+
+        x = uninstantiated_transformers[self.transformer_name](
+            copy=True,
+            **minimal_attribute_dict[self.transformer_name],
+        )
+
+        x = x.fit(df, df["c"])
+
+        df_transformed = x.transform(df)
+
+        assert df_transformed is not df
+
+    def test_no_rows_error(self, instantiated_transformers):
+        """Test an error is raised if X has no rows."""
+        df = d.create_df_10()
+
+        x = instantiated_transformers[self.transformer_name]
+
+        x = x.fit(df, df["c"])
+
+        df = pd.DataFrame(columns=["a", "b", "c"])
+
+        with pytest.raises(
+            ValueError,
+            match=re.escape(f"{self.transformer_name}: X has no rows; (0, 3)"),
+        ):
+            x.transform(df)
+
+
+class ColumnsCheckTests:
+    """
+    Tests for columns_check method.
+    Note this deliberately avoids starting with "Tests" so that the tests are not run on import.
+    """
+
+    def test_non_pd_df_error(
+        self,
+        instantiated_transformers,
+    ):
+        """Test an error is raised if X is not passed as a pd.DataFrame."""
+
+        x = instantiated_transformers[self.transformer_name]
+
+        with pytest.raises(
+            TypeError,
+            match=f"{self.transformer_name}: X should be a pd.DataFrame",
+        ):
+            x.columns_check(X=[1, 2, 3, 4, 5, 6])
+
+    @pytest.mark.parametrize("non_list", [1, True, {"a": 1}, None, "True"])
+    def test_columns_not_list_error(
+        self,
+        non_list,
+        instantiated_transformers,
+    ):
+        """Test an error is raised if self.columns is not a list."""
+        df = d.create_df_1()
+
+        x = instantiated_transformers[self.transformer_name]
+
+        x.columns = non_list
+
+        with pytest.raises(
+            TypeError,
+            match=f"{self.transformer_name}: self.columns should be a list",
+        ):
+            x.columns_check(X=df)
+
+    def test_columns_not_in_X_error(
+        self,
+        instantiated_transformers,
+    ):
+        """Test an error is raised if self.columns contains a value not in X."""
+        df = d.create_df_1()
+
+        x = instantiated_transformers[self.transformer_name]
+
+        x.columns = ["a", "z"]
+
+        with pytest.raises(ValueError):
+            x.columns_check(X=df)
+
+
+class CombineXYTests:
+    """
+    Tests for the BaseTransformer._combine_X_y method.
+    Note this deliberately avoids starting with "Tests" so that the tests are not run on import.
+    """
+
+    @pytest.mark.parametrize("non_df", [1, True, "a", [1, 2], {"a": 1}, None])
+    def test_X_not_DataFrame_error(
+        self,
+        non_df,
+        instantiated_transformers,
+    ):
+        """Test an exception is raised if X is not a pd.DataFrame."""
+
+        x = instantiated_transformers[self.transformer_name]
+
+        with pytest.raises(
+            TypeError,
+            match=f"{self.transformer_name}: X should be a pd.DataFrame",
+        ):
+            x._combine_X_y(X=non_df, y=pd.Series([1, 2]))
+
+    @pytest.mark.parametrize("non_series", [1, True, "a", [1, 2], {"a": 1}])
+    def test_y_not_Series_error(
+        self,
+        non_series,
+        instantiated_transformers,
+    ):
+        """Test an exception is raised if y is not a pd.Series."""
+
+        x = instantiated_transformers[self.transformer_name]
+
+        with pytest.raises(
+            TypeError,
+            match=f"{self.transformer_name}: y should be a pd.Series",
+        ):
+            x._combine_X_y(X=pd.DataFrame({"a": [1, 2]}), y=non_series)
+
+    def test_X_and_y_different_number_of_rows_error(
+        self,
+        instantiated_transformers,
+    ):
+        """Test an exception is raised if X and y have different numbers of rows."""
+
+        x = instantiated_transformers[self.transformer_name]
+
+        with pytest.raises(
+            ValueError,
+            match=re.escape(
+                f"{self.transformer_name}: X and y have different numbers of rows (2 vs 1)",
+            ),
+        ):
+            x._combine_X_y(X=pd.DataFrame({"a": [1, 2]}), y=pd.Series([2]))
+
+    def test_X_and_y_different_indexes_warning(
+        self,
+        instantiated_transformers,
+    ):
+        """Test a warning is raised if X and y have different indexes, but the output is still X and y."""
+
+        x = instantiated_transformers[self.transformer_name]
+
+        with pytest.warns(
+            UserWarning,
+            match=f"{self.transformer_name}: X and y do not have equal indexes",
+        ):
+            x._combine_X_y(
+                X=pd.DataFrame({"a": [1, 2]}, index=[1, 2]),
+                y=pd.Series([2, 4]),
+            )
+
+    def test_output_same_indexes(
+        self,
+        instantiated_transformers,
+    ):
+        """Test output is correct if X and y have the same index."""
+        x = instantiated_transformers[self.transformer_name]
+
+        result = x._combine_X_y(
+            X=pd.DataFrame({"a": [1, 2]}, index=[1, 2]),
+            y=pd.Series([2, 4], index=[1, 2]),
+        )
+
+        expected_output = pd.DataFrame(
+            {"a": [1, 2], "_temporary_response": [2, 4]},
+            index=[1, 2],
+        )
+
+        pd.testing.assert_frame_equal(result, expected_output)
+
+
+class CheckWeightsColumnTests:
+    """
+    tests for check_weights_column method.
+    Note this deliberately avoids starting with "Tests" so that the tests are not run on import.
+    """
+
+    def get_df_error_combos():
+        return [
+            (
+                pd.DataFrame({"a": [1, 2], "b": [3, 4]}),
+                r"weight col \(c\) is not present in columns of data",
+                "c",
+            ),
+            (
+                pd.DataFrame({"a": [1, 2], "b": ["a", "b"]}),
+                r"weight column must be numeric.",
+                "b",
+            ),
+            (
+                pd.DataFrame({"a": [1, 2], "b": [-1, 0]}),
+                r"weight column must be positive",
+                "b",
+            ),
+            (
+                pd.DataFrame({"a": [1, 2], "b": [np.NaN, 0]}),
+                r"weight column must be non-null",
+                "b",
+            ),
+        ]
+
+    @pytest.mark.parametrize("df, error, col", get_df_error_combos())
+    def test_weight_not_in_X_error(self, df, error, col, uninstantiated_transformers):
+        """Test an error is raised if weight is not in X."""
+
+        with pytest.raises(
+            ValueError,
+            match=error,
+        ):
+            uninstantiated_transformers[self.transformer_name].check_weights_column(
+                df,
+                col,
+            )
+
+
+class OtherBaseBehaviourTests(
+    ColumnsCheckTests,
+    CheckWeightsColumnTests,
+    CombineXYTests,
+):
+    """
+    Class to collect and hold tests for BaseTransformerBehaviour outside the three standard methods.
+    Note this deliberately avoids starting with "Tests" so that the tests are not run on import.
+    """
